requirements.txt
Django==4.1.7
django-debug-toolbar==4.0.0
djangorestframework==3.14.0

pyyaml==6
uritemplate==4.1.1

pillow~=9.3.0

pylint==2.17.4